// frontend/types.ts
export interface Image {
  id: number;
  image_url: string;
  alt_text: string;
  caption?: string;
  show_on_home?: boolean;
  image_type: string;
  linked_object_type: string | null;
}
// ...
export interface Education {
  id: number;
  institution: string;
  degree: string;
  field_of_study: string;
  start_date: string;
  end_date?: string;
  is_current: boolean;
  logo: string;
  description: string;
  slug: string;
  images?: Image[];
  show_on_home?: boolean;
}